Code divisions to make understanding separate.
